+++ packages/nativescript-material-bottomsheet/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.9](https://github.com/Akylas/nativescript-material-components/compare/v3.1.8...v3.1.9) (2020-02-24)
+
+**Note:** Version bump only for package nativescript-material-bottomsheet
+
+
+
+
+
+## [3.1.8](https://github.com/Akylas/nativescript-material-components/compare/v3.1.7...v3.1.8) (2020-02-20)
+
+**Note:** Version bump only for package nativescript-material-bottomsheet
+
+
+
+
+
 ## [3.1.7](https://github.com/Akylas/nativescript-material-components/compare/v3.1.6...v3.1.7) (2020-02-18)
 
 **Note:** Version bump only for package nativescript-material-bottomsheet